addressed PR feedbacks: 1. renamed TokenRefresher to TokenProvider 1. renamed Refresh() function to Acquire() function 1. error message update
