+++ auth/providers/azure/graph/clientcredential_tokenprovider_test.go
@@ -18,11 +18,10 @@ package graph
 import (
 	"fmt"
 	"net/http"
-	"net/http/httptest"
 	"testing"
 )
 
-func TestOBOTokenRefresher(t *testing.T) {
+func TestClientCredentialTokenRefresher(t *testing.T) {
 	const (
 		inputAccessToken    = "inputAccessToken"
 		oboAccessToken      = "oboAccessToken"
@@ -31,8 +30,7 @@ func TestOBOTokenRefresher(t *testing.T) {
 		scope               = "https://graph.microsoft.com/.default"
 		oboResponse         = `{"token_type":"Bearer","expires_in":3599,"access_token":"%s"}`
 		expectedContentType = "application/x-www-form-urlencoded"
-		expectedTokenUse    = "on_behalf_of"
-		expectedGrantType   = "urn:ietf:params:oauth:grant-type:jwt-bearer"
+		expectedGrantType   = "client_credentials"
 		expectedTokneType   = "Bearer"
 	)
 
@@ -50,12 +48,6 @@ func TestOBOTokenRefresher(t *testing.T) {
 			if req.FormValue("client_secret") != clientSecret {
 				t.Errorf("expected client_secret: %s, actual: %s", clientSecret, req.FormValue("client_secret"))
 			}
-			if req.FormValue("assertion") != inputAccessToken {
-				t.Errorf("expected assertion: %s, actual: %s", inputAccessToken, req.FormValue("assertion"))
-			}
-			if req.FormValue("requested_token_use") != expectedTokenUse {
-				t.Errorf("expected requested_token_use: %s, actual: %s", expectedTokenUse, req.FormValue("requested_token_use"))
-			}
 			if req.FormValue("scope") != scope {
 				t.Errorf("expected scope: %s, actual: %s", scope, req.FormValue("scope"))
 			}
@@ -67,8 +59,8 @@ func TestOBOTokenRefresher(t *testing.T) {
 
 		defer stopTestServer(t, s)
 
-		r := NewOBOTokenRefresher(clientID, clientSecret, s.URL, scope)
-		resp, err := r.Refresh(inputAccessToken)
+		r := NewClientCredentialTokenRefresher(clientID, clientSecret, s.URL, scope)
+		resp, err := r.Acquire(inputAccessToken)
 		if err != nil {
 			t.Fatalf("refresh should not return error: %s", err)
 		}
@@ -95,12 +87,6 @@ func TestOBOTokenRefresher(t *testing.T) {
 			if req.FormValue("client_secret") != clientSecret {
 				t.Errorf("expected client_secret: %s, actual: %s", clientSecret, req.FormValue("client_secret"))
 			}
-			if req.FormValue("assertion") != inputAccessToken {
-				t.Errorf("expected assertion: %s, actual: %s", inputAccessToken, req.FormValue("assertion"))
-			}
-			if req.FormValue("requested_token_use") != expectedTokenUse {
-				t.Errorf("expected requested_token_use: %s, actual: %s", expectedTokenUse, req.FormValue("requested_token_use"))
-			}
 			if req.FormValue("scope") != scope {
 				t.Errorf("expected scope: %s, actual: %s", scope, req.FormValue("scope"))
 			}
@@ -114,8 +100,8 @@ func TestOBOTokenRefresher(t *testing.T) {
 
 		defer stopTestServer(t, s)
 
-		r := NewOBOTokenRefresher(clientID, clientSecret, s.URL, scope)
-		resp, err := r.Refresh(inputAccessToken)
+		r := NewClientCredentialTokenRefresher(clientID, clientSecret, s.URL, scope)
+		resp, err := r.Acquire(inputAccessToken)
 		if err == nil {
 			t.Error("refresh should return error")
 		}
@@ -127,14 +113,4 @@ func TestOBOTokenRefresher(t *testing.T) {
 			t.Errorf("expected token type: %s should be empty", resp.TokenType)
 		}
 	})
-}
-
-func startTestServer(t *testing.T, handler func(rw http.ResponseWriter, req *http.Request)) *httptest.Server {
-	t.Helper()
-	return httptest.NewServer(http.HandlerFunc(handler))
-}
-
-func stopTestServer(t *testing.T, server *httptest.Server) {
-	t.Helper()
-	server.Close()
 }
